Update method for generating order months dropdown
In #55510 we changed how the Months filter dropdown on the Orders list
table is generated so that it uses a much more efficient query. However,
it introduced a bug that could happen at the end of the month when the
following month had fewer days in it. This was surfaced via a unit test,
fortunately. In this PR we are making the method more robust so that the
number of days in the month doesn't matter, and also improving the
behavior during an edge case where all existing orders are in the future