Simplify months query and option building algo
This does away with conditionally running two separate queries to get
the minimum and maximum order creation dates, in favor of one query that
selects the min and max and returns them as a single row. It also makes
some other changes to try and be consistent about the timezones of the
dates involved.